refactor export entry function and file locations
